Update verbose mode to display shell commands
`waf build -v` used to display runner output as a python array that is
used as input to 'exec'. This is inconvenient to users who might want to
run the commands directly on the shell for debugging purposes. This
commit changes the verbose output to show a command that can be directly
executed in a shell. It does this using a simple heuristic, and can't
handle embedded quotes.  To get the previous verbose behavior, use `waf
build --zones runner_real` instead.