Treat clang as a separate compiler in UPLID.
This change stops overriding *all* versions of clang to be reported
as `gcc-clang`, as future versions of LLVM may introduce different
C++1y features in specific versions.  The build system needs to be
able to differentiate between the version of LLVM being used.

Renames all `-gcc-clang` rules in `default.opts` to `-clang` and
duplicates all `-gcc` rules for Linux and Darwin to seed the `-clang`
rules. (It is assumed AIX and Solaris are not supported for use with
LLVM, so platform-specific `-gcc` rules on those platforms are not
copied.)