Refactor code and add new features
Code Refactor
-------------
The existing code has been refactored to enable easier maintenance and
reuse.  The python packages are now loacted in the bin/lib directory,
which currently contains 3 packages "bdebld", "legacy", and "tests".

The "legacy" package contains just a single source file that preserves backwards
compatibliity with the older version of the wscript.

The "bdebld" pacakge contains code to build BDE-style repositories, and
it contains 3 sub-packages: meta, waf, and runtest.

The "meta" sub-package does not depend on waf. It's purpose is to load
the repository structure into a "build configuration", which contains
the necessary structures and build flags to build a repo.

The "waf" sub-package adapts a "build configuration" so that it can be
built with waf.

The "runtests" sub-package provides facilities to run a BDE-style test
driver in a configurable way.

The "tests" package contains all test drivers.  A shell script
"run_tests.py" has been added to bin/lib that runs all of the test
drivers.  If the "-c" option is specified, then "run_tests.py" will
collect code coverage data using coverage.py, and once this is done, you
can use the commadn "coverage report" to view code coverage statistics.

New features
------------

- "waf build" now builds in topological order (simpler debugging)

- "waf configure --verify"
  checks for cyclic dependencies in a repo

- "waf graph" draws dot dependency diagrams
  E.g.,
  waf configure
  waf graph # draw all UORs
  waf graph --target bsl # draw packages in bsl
  waf graph --target bslalg # draw components in bslalg