Remove the extracted waf source files
The waf python script packs some source files that get extracted when
the script is first run.  We decided to check the extracted source files
into this repo because it gets installed to a read-only location on some
of our development machines, and the extraction process used by the
script would fail.  Unfortunately, this means that on windows or when
python 3.x used, the waf source files checked into the repo are unused,
because they are specific to python 2.x.

We have sinced solved the issue on our development machines using a
better install process, so the the waf source files are no longer
required to be checked in.