Support alternate application source file name
If the application package name starts with "m_", the main function can
be defined in either "<package_name>.m.cpp" or the same except with the
"m_" prefix removed.  For example, the main source file of an
application package "m_app" can be either "m_app.m.cpp" or "app.m.cpp".