Fix parsing .bdelayoutconfig in python 2
The previous commit fixed the problem with using .bdelayoutconfig on
python 3 but broke the feature on python 2.

It turns out that the Node interface in waf expects values of type
string, and JSON parser from python 2 stores the values of the resulting
dictionary in type Unicode, which needs to be encoded to a string.