Support empty pub files
We need to distinguish between the case when the pub file does not exist
and the case when the pub file exist but is empty.  If the pub file does
not exist, then every header file should be exported; if the pub file
does exist but is empty, then no header file should be exported.