Use a canonical Python3 version
The current check differs from file to file, some require
python, some require python3.8. The current version of
python is 3.10, and it's backward compatible, so we only
need to test if a python version is *below* python3.8 before
bailing out.

This future proofs the code and allows linux distros and
windows installations to use newer python versions.